Search Bar Implemented
Search bar is now implemented inside of the AllPosts element as a functional component. We filter based on name of the film by mapping the filterItems() function which returns a new array of posts.

Next I will need to implement a method which allows us to search without needing the search terms to be case sensitive